Refactored google cloud storage utils
