+++ server/utils/googleCloudStorageUtils.js
